Refactor LED interfacing and build infra
This commit factors out the LED-specific code into its own file, as
part of a process of getting the codebase into a state that better
matches the argali ideal.

This commit also makes a bunch of changes to the build
infrastructure.  The big motivation was to support the newly-separated
`leds.c` code, as we want to be able to compile it in on Linux, but
without hardware.  It turned out that the Unity Test makefile was
stomping on some variables (like, oh, `$(CFLAGS)`) and causing `TEST`
to be defined everywhere.  This rejiggering of the Makefiles was to
support both building for hardware and on the host side.

As part of this, I wound up chasing my tail for most of a day due to
both this problem as well as a really dumb thinko that was hard to
debug.  On the positive side, this commit therefore begins the process
of supporting both gdbgui and svd-dump, which is decent enough.