Add gdbgui and svd-dump
I got very confused about a few things, which lead me down a debugging
rabbit hole.  As part of that, I installed a few handy tools, which
are now going into the repository with this commit.

gdbgui is a really neat web-based interface to `gdb`.  It provides an
interface a lot like the (excellent) debugger in Visual Studio.
There's also the ability to plot values over time, much like VisualGDB
allows you to do in VS.  But without requiring Visual Studio, or
Windows, or making it impossible to use command-line debuggers.  It's
great.

svd-dump is a random python script from a gist that wraps up a lot of
the annoying bits of pulling registers into less awful gdb commands.
That said, it's very rough around the edges, and I think there's a lot
of low-hanging fruit to improve it.  Sadly, it does not support full
register decodes (another thing VisualGDB does pretty well).