console: bugfix and reformat to use macros
The main notivation for this commit is to move over to using macros
for all the resources used in the console code.  This is best practice
so you don't accidentally use the wrong pin in two different places.

This commit fixes a dumb buglet in the serial console code, where we
set up pin D7 to be our RX line via AFs, but enabled pin D9 instead of
D7.

Best Practices: because you should be making *new* stupid mistakes,
not repeating old ones!(TM)