docker_shell.sh: mount the project directory in the docker container
We currently mount `argali/code/` as `/code` in the docker container.
This works great right up until we need to use the git command line to
get version information: we don't have access to `.git/` from
`/code`.  Therefore, we add a second volume mount, so `argali/` is
mounted in the container (and we can therefore run `git` commands).