Adds openocd build
So it turns out that the version of OpenOCD included in ubuntu repo is
rather old.  This is because OpenOCD doesn't release very often, but
it means that some relatively common chips aren't supported
out-of-the-box on Ubuntu.

We therefore do a checkout of a "known good" version of OpenOCD (for
our chips of interest, at least...) and build it.