CHANGE PLACES!  Restructure directories to support multi-target
This is a big commit in terms of the stuff it touches, but it doesn't
actually do terribly much.  Specifically, it only shuffles files
around so that we can more easily support multiple concurrent
hardware-specific files at once.

That shuffling results in a lot of files moving around, and a new
directory structure being in place.  However, those directories are
pretty straightforward, and the moves were primarily just `git mv`.
That said, I also factored out the few remaining places where
hardware-specific includes or code was in the top-level src/
directory.

These changes build both the unit tests and the firmware for the F767,
and I have already built and run firmware for the F413 nucleo from it.
This is a handy target, as it's very similar to the F767 (pretty sure
ST used the same IP blocks for most of the peripherals...).

I'll be committing the F413 code shortly, and then need to document
the process of adding a new platform (by adding a nucleo F303).

There are also a few ugly corners left (ELF files being left in code/
instead of in the appropriate build directory, the .gdbinit is still
specialized for the F767, and doxygen might be a bit hosed at the
moment).

But this is a big step towards the goal of demonstrating a way to
support multiple PCBs/platforms from the same repo without too much
stepping on each other's toes.