targets/nucleo_f413zh: adds support for nucleo F413ZH
This commit adds the files needed to support the Nucleo STM32F413ZH
evaluation board.  There are several problematic pieces to it, but
this is the proof-of-concept to get the ball rolling.

It can be built with:

make TARGET=nucleo_f413zh docker-build

and then you can program it via the openocd daemon and gdbgui
containers with

make TARGET=nucleo_f413zh docker-openocd-daemon

and then, in a separate terminal,

make docker-gdbgui

And use `load` from gdb to load the final ELF file.  (You can also run
the `make TARGET=nucleo_f413zh` from within the gdb console there.)