Makefiles: don't build binaries in code/
I believe this completes the multi-platform build and debug
capability.  This moves the .elf file down into the appropriate build
subdirectory.

There were also places where 'make' running inside of docker was
missing context that 'make' on the host OS had (eg: where the output
binary was located).  This commit reworks how we pass state into the
docker container, which seems to work (but is ugly).