Packet handling: adds an encapsulated serial protocol framing
This commit adds a few files to build HDLC-ish framed packets that we
can use over the serial port.  There is a corresponding python
implementation, which we will tie together to let the test stand code
interact directly with both the DUT and the test rig.

This is not yet wired up to the actual serial handler, as it interacts
with the logging code in a meaningful way.  That's the next step,
along with tightening up the python side.