WIP F413ZH DAC: Adds very rudimentary DAC driver for F413ZH board
It seems like TamoDevBoard should make some annoying sound when it's
bored, so we add a quick and dirty DAC interface to let it go eeeee at
users or whatever.

This is the first pass of getting it going based on the libopencm3
example, which is very rudimentary and rough around the edges.  We
will smooth this over, then implement the F7 version before merging
back to the main repo.