DAC/ADC: Adds WIP (but working) DAC and ADC drivers for STM32F4/F7
This is the first diff to get both DAC and ADC working on the STM32F4
and STM32F7 (and likely many, many others, based on ST docs).

In the current kludged `main()`, if you push the button, it will dump
the ADC buffer on the console.  You can run a wire from the DAC output
to the ADC input (PA4-PA0, second from bottom on the left of CN7 to
the third from bottom on left of CN10) to sniff the data directly.

This is a good proof-of-concept of signal generation at a desired
frequency, along with simple ADC.

Lots of work to go: ADC needs clock control, and we need to be able to
turn the DAC on and off, etc etc.

Note that the code between the F4 and F7 is quite literally identical;
the only difference is in the comments.  It may make sense to factor
this out into an stm32_common directory in targets/, and add
provisions in there to reference the appropriate docs from ST.