sin_gen: Adds a scaling component
While testing the divider logic, I noticed that higher frequencies
would get really terrible waveforms as output.  I thought this was a
bandwidth limit at first, but came across AN4566, "Extending the DAC
performance of STM32 microcontrollers" while looking for proper specs
on the bandwidth.  It mentioned slew rate as a consideration in op
amps, and that made me wonder if a smaller amplitude waveform would
actually work at the higher frequencies.

Lo and behold, it did.  At this point, it may also make sense to add
an offset value as well, so you can center your waveform somewhere
besides 127.  For now, I'm going to leave it as-is, but that's a
straightforward change that parallels this one closely.