F413ZH ADC WIP Checkpoint
ADC now has a configurable frequency, insofar as that's possible on
the STM32s.  It will give you the best effort at the frequency you
specify.

The pause and unpause ADC functionality seems to mostly work, but I
need to add more functionality to main to test it (pause, then print
twice).

adc_pause now returns the offset from the tail of the buffer that your
current transaction was interrupted.  main uses that to wrap the data
dump nicely.

main also prints out the data in a much more pleasant format to
eyeball.

F767 implementation to follow, and I need to actually figure out
if I want to keep this ADC free-running clock, or if I want to tie it
into the TIM2 TRGO like the DAC.  We could then sample on the falling
edge, which would let the DAC value get updated and then immediately
take a reading, without missing a beat.  That'd be neat.

APIs need to be stared at, as well.  I may want to pull more stuff
from parameters into structs, especially on the DAC side.  It makes
things a bit more HAL-like; while there are many things I dislike
about HAL, it's at least a lot easier to push out a coherent state in
a few lines of code.  There's just a giant whap of struct building
above it.