sin_gen: add a blind buffer fill
This commit adds the ability to blithely fill a buffer all the way,
with no regard for whether or it makes particular sense.

This is handy if you want to generate two tones of different
frequencies with the same buffer length (for, say, DTMF).

It cannot be overstated how nasty of a solution this is.  It's going
to cause all sorts of ugliness in the output signal, but it might just
work, so we're trying it.

It also seems like it might be useful if someone really knows what
they're doing and sets the buflen appropriately in their request.