TamoDevBoard recites digits of pi via DTMF when bored
This diff adds the ability for TamoDevBoard to recite the digits of pi
when bored.  It does this by putting DTMF out on the DAC, and using
the ADC to decode those digits and confirm them.

This only supports the F413ZH target at the moment, 767ZI support will
be coming as soon as the F413ZH stabilizes.  There are some more
changes to the ADC and DAC APIs that would be nice, and possibly
factoring out DMA and Timer drivers.  This will simplify the code in
the ADC and DAC code, and reduce duplication of the setup of those
peripherals.