DTMF Detection: move to target-provided ADC Prescalers
The F767ZI was having a very hard time decoding tones given the
combination of buffer length and sampling rate it was able to hit with
the default prescaler/period combination.

This commit moves things around so that it can be 8kHz on all
platforms, with each platform supplying the required prescaler values
required to achieve that.