F767ZI DAC: Fixes start/stop, adds workaround for erratum
There was a stupid bug in the dac_stop()/dac_start() previously, where
we weren't cleanly disabling the DMA connection. That is now fixed.

While looking for other possible problems (as part of assembling the
README.md for this target), I found a possibly relevant erratum that
could have explained some DTMF decode problems.  It turns out that
those were actually problems with the sampling rate, fixed in
6163ccd66e90b182f2f5a29aad5ad97ac04b9441.  However, it seems prudent
to leave the workaround logging in place, so we can detect it if it
happens.