Bugfix for sin_gen test case
There was a segfault in one sin_gen test case, which had gone
unnoticed until I was testing this packet code.  This fixes
it (it was memset'ing against a pointer that hadn't been filled in
yet; I imagine that it was randomly still pointing at the single test
buffer from the last test until now).