Framing implementation: quote everywhere
There were a few edge cases that I found while testing the framing
implementation that I didn't like.  In particular, it was possible for
a ~ to wind up in the header fields, which was ambiguous as to whether
it was a value in the current frame or a preamble for a new
frame (after the current frame was abandoned).

This moves things around a bit to escape everything but the actual
flags, which makes the code a bit simpler, and allows us to treat any
unescaped flag as a reset for the line.  This should make things
significantly more robust in the face of partial transmissions.