ADC Configuration rework, adds ADC read command to EOL console
This adds the ability to read data off the ADC in the EOL console.  It
refactors the ADC configuration interface to enable this, as well as
making it possible to get data from multiple inputs and in either 8 or
12b resolution (for EOL station applications).

This is yet another F413-specific diff.  It needs to be ported to the
F767, but I will work on polishing this one further first.