Add docker documentation, small tweaks for forking argali
This diff adds documentation on how to use docker effectively.

This diff also adds a few niceties for forking argali off into a
different named docker image.  This is helpful if you want to take
argali and develop a product with it, without intending to upstream
any of your changes.  That kind of use is exactly what is intended by
our CC0 licensing, so it's good to support better.