Updates python unit tests, moves to unified protobuf file
This is the last big step on the python side.  It moves to a unified
protobuf definition file.  We will put all the packets into a single
.proto file, to allow multiplexing packets without adding headers of
our own.  We could use HDLC here, but the plan is to support network
packets as well as bluetooth at some point, so we won't have those
HDLC-specific fields to work with.

This adds a unit test that round-trips through a janky version of the
parser, which proves out that it works.

Next is to get the microcontroller side using this to log things, and
we're good to go on rebuilding the tethered EOL firmware with the new
packet formats.