Completes removal of nanopb
Nanopb wound up adding far too much juggling and overhead to what
should be very straightforward/trivial wire protocols.  Most
significantly, it required either external framing of packet
type (which defeats half the point) or a weird spaghetti of "partial
decode, which calls a callback so you can register the appropriate
callbacks for the rest of the packet" thing.  This meant that adding a
new packet resulted in a bunch of little diffs, some of which were
inside-out callback-to-set-up-callback logic.  That is more cognitive
overhead than I would expect anyone to track with something as
infrequent as packet addition.

I may streamline the packet parsing code in the future, though, as
it's currently ugly.  It's pretty easy to imagine the python code
generating C parsing code and a table of packet registrations, which
the embedded code then runs.  This also makes it easy to run unit
tests on the generated code, so we can verify and validate it very
thoroughly.