Checkpoint on the way to full command line control
This adds a bunch of packet generation stuff to the argali_tether
library, which is currently mostly-compatible with the existing
protocol.  Note that packet parsing hasn't been tied into the
ArgaliTarget class yet, so that still needs to be done.  However, it's
in pretty good shape thus far, and the command line tools are
sufficient for controlling the DAC and fetching ADC samples.  This is
a lot of the most sophisticated EOL functionality that isn't straight
up application-specific state machine testing, so it's good to get
them nailed down.

There are also some nice helpers for turning packets into command line
tools in their own right, which is pretty handy.