Updates to ADC to support 4 channel input
I'm working on a separate project using argali, and wanted to have
more channels of input simultaneously.  I also wanted better control
of the inter-sample phase shifts for audio processing, so I made that
a parameter.  The deeper sampling of a microphone array required a
much longer sample buffer, which is added.

The rest of this is bugfixes and other things to make life a bit nicer
when using argali's tethered mode from outside an argali clone repo.